t/op/utfhash.t: Fix to work on EBCDIC 1047
This .t thought it could have an __END__ in the middle of DATA input,
and the first read would read in up to that, and the second would read
the rest.  In bisecting, I couldn't find a time when this ever worked.
Now this has a marker, and does a split on that marker, choosing the
first or second half depending.

Note that this only works on ASCII and EBCDIC 1047 platforms.  It could
be extended for the other code pages Perl purportedly supports.