Generalize NBSP and SHY handling for non-ASCII platforms
The No-Break Space and Soft Hyphen are used in 6 modules.  This
generalizes so they can be handled fully on non-ASCII platforms.  A
recent patch had already fixed this this for one area of code, but it
turns out that they are used in more than one place.  In most of those
places, they were handled somewhat gracefully for non-ASCII platforms,
but this patch makes them work completely correctly.

I used global scalar variables in the base module to store what the
native characters are for these code points, as the calculation of
what they should be is not obvious, and so should be done in a single
place.  An unlikely pitfall is that these scalars are not read-only; I
suppose a subroutine could be used instead, I suppose, but I thought
that this was adequate.