Fix some escapes to work on non-ASCII platforms
This same code is repeated in multiple places.  I chose to not
consolidate it.  The comments indicate that it was known it would work
only on ASCII, but since v5.8, there is the capability to make it easily
working on non-ASCII as well, using the translation functions available
starting in that release